# docs/specs/openapi.yaml
openapi: 3.1.0
info:
  title: Euro Intermed B2B Platform API
  version: 1.0.0
  description: >
    REST + real-time API for the Euro Intermed B2B lead-qualification platform
    (Angrosist + PalletClearance, Phase 1; SkalYou, Phase 2). See the companion
    human-readable contract in docs/specs/API_CONTRACT.md for conventions, auth,
    error envelope, pagination, rate limits, validation, and the SSE/WebSocket
    streaming contract (which is described there rather than fully modeled here).
    Phase tags appear in operation summaries: [P1], [P1-demo], [P2].
  contact:
    name: Euro Intermed Platform
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: '{baseUrl}/api'
    description: >
      Base URL is environment-provided and never hardcoded. Frontend uses
      VITE_API_URL (Vite). Production is Cloud Run behind Cloudflare; the demo is
      a Vercel deployment.
    variables:
      baseUrl:
        default: https://api.example.com
...
security: []
paths:
...
  /leads:
    get:
      tags: [dashboard]
      summary: '[P1] List leads (filtered, paginated)'
      description: >
        Authenticated lead list with filters and cursor pagination. NOTE: the
        shipped demo returns a bare LeadSummary[] without auth or paging; this
        operation documents the Phase-1 target (paginated envelope). The
        response schema below accepts either shape so the spec matches current
        and target reality.
      operationId: listLeads
      security:
        - staffJwt: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/LeadStatus' }
        - name: vertical
          in: query
          schema: { $ref: '#/components/schemas/Vertical' }
        - name: assigned_to
          in: query
          schema: { type: string }
        - name: country
          in: query
          description: ISO 3166-1 alpha-2.
          schema: { type: string, minLength: 2, maxLength: 2 }
        - name: market
          in: query
          schema: { type: string }
        - name: search
          in: query
          description: Full-text over company / product / summary.
          schema: { type: string, maxLength: 200 }
      responses:
        '200':
          description: Leads page (envelope) or bare array (demo).
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
            RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LeadListPage'
                  - type: array
                    items: { $ref: '#/components/schemas/LeadSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /leads/{id}:
    get:
      tags: [dashboard]
      summary: '[P1] Lead detail (transcript + extracted + typed request)'
      operationId: getLead
      security:
        - staffJwt: []
      parameters:
        - $ref: '#/components/parameters/LeadId'
      responses:
        '200':
          description: Lead detail.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LeadDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFoundLegacy' }
  /leads/{id}/status:
    patch:
      tags: [dashboard]
      summary: '[P1] Update lead/offer status'
      description: >
        Manual offer tracking. status must be a valid transition in the lead
        state machine; illegal transition returns 409. Writes activity_logs.
      operationId: updateLeadStatus
      security:
        - staffJwt: []
      parameters:
        - $ref: '#/components/parameters/LeadId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { $ref: '#/components/schemas/LeadStatus' }
                value:
                  type: number
                  minimum: 0
                  description: Offer/pipeline value (optional).
                note: { type: string, maxLength: 2000 }
              required: [status]
      responses:
        '200':
          description: Updated lead.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LeadSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /leads/{id}/offer:
    patch:
      tags: [dashboard]
      summary: '[P1] Update manual offer tracking (alias of /leads/{id}/status)'
      description: >
        Manual offer tracking. Sets leads.status / offer_value / offer_note and
        writes an offer.update activity_logs row. status is validated against the
        lead_statuses lookup; value must be >= 0; note <= 2000 chars. At least one
        of status/value/note is required. Same handler as PATCH /leads/{id}/status.
      operationId: updateLeadOffer
      security:
        - staffJwt: []
      parameters:
        - $ref: '#/components/parameters/LeadId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { $ref: '#/components/schemas/LeadStatus' }
                value:
                  type: number
                  minimum: 0
                  description: Offer/pipeline value (optional).
                note: { type: string, maxLength: 2000 }
      responses:
        '200':
          description: Updated lead.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LeadSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /leads/{id}/assign:
    post:
      tags: [dashboard]
      summary: '[P1] Assign / unassign a lead'
      description: >
        Sets leads.assigned_to (user_id, or null to unassign) and writes a
        lead.assign activity_logs row. The target user must exist when non-null.
      operationId: assignLead
      security:
        - staffJwt: []
      parameters:
        - $ref: '#/components/parameters/LeadId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: [string, 'null']
                  description: User id to assign, or null to unassign.
              required: [user_id]
      responses:
        '200':
          description: Updated lead.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LeadSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /handoffs:
    get:
      tags: [dashboard]
      summary: '[P1] Handoff queue (needs_human leads)'
      description: >
        Leads with needs_human=true, newest first, each carrying a short snippet
        of the conversation's last message. Also reachable at /handoff (alias).
      operationId: listHandoffs
      security:
        - staffJwt: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Handoff queue page.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/HandoffItem' }
                  page: { $ref: '#/components/schemas/PageInfo' }
                required: [data, page]
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /companies:
    get:
      tags: [dashboard]
      summary: '[P1] B2B directory list'
      operationId: listCompanies
      security:
        - staffJwt: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: roles
          in: query
          description: Match companies having any of these roles.
          style: form
          explode: true
          schema:
            type: array
            items: { $ref: '#/components/schemas/CompanyRole' }
        - name: country
          in: query
          schema: { type: string, minLength: 2, maxLength: 2 }
        - name: vat_status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string, maxLength: 200 }
      responses:
        '200':
          description: Companies page.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Company' }
                  page: { $ref: '#/components/schemas/PageInfo' }
                required: [data, page]
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /companies/{id}:
    get:
      tags: [dashboard]
      summary: '[P1] Company detail'
      operationId: getCompany
      security:
        - staffJwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Company detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CompanyDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /kpis:
    get:
      tags: [dashboard]
      summary: '[P1] Dashboard KPIs'
      operationId: getKpis
      security:
        - staffJwt: []
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: KPI aggregates.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Kpis' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    staffJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Staff/admin session JWT (claims: sub, role, exp).'
    providerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '[P2] Provider JWT from Google or email OTP (scoped to provider_id).'
    clientJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '[P2] Short-lived client JWT exchanged from a magic-link token.'
  headers:
    XRequestId:
      description: Correlation id, echoed or server-generated.
      schema: { type: string }
    RateLimitLimit:
      description: Requests allowed in the current window.
      schema: { type: integer }
    RateLimitRemaining:
      description: Requests remaining in the current window.
      schema: { type: integer }
    RateLimitReset:
      description: Seconds until the window resets.
      schema: { type: integer }
...
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size (default 25, max 100; out-of-range is clamped).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from the previous page's next_cursor.
      schema: { type: string }
    LeadId:
      name: id
      in: path
      required: true
      description: Lead id.
      schema:
        type: string
        minLength: 1
  responses:
    BadRequest:
      description: Validation failure / malformed request.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing/invalid/expired credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Authenticated but not allowed (role/ownership/signature).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: Invalid state transition.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Unprocessable:
      description: Semantically invalid (e.g. unknown enum value).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    NotFoundLegacy:
      description: Not found (legacy error shape on demo handlers).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/LegacyError' }
...
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stable machine-readable code (SCREAMING_SNAKE_CASE).
              examples: [VALIDATION_FAILED, UNAUTHENTICATED, FORBIDDEN, NOT_FOUND, CONFLICT, EXPIRED, UNPROCESSABLE, RATE_LIMITED, INTERNAL, UPSTREAM_UNAVAILABLE]
            message:
              type: string
              description: Human-readable, safe to surface.
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }
                required: [field, issue]
            request_id: { type: string }
          required: [code, message]
      required: [error]
    LegacyError:
      type: object
      description: Legacy demo error shape (already shipped). New code uses ErrorEnvelope.
      properties:
        error: { type: string }
      required: [error]
    PageInfo:
      type: object
      properties:
        next_cursor:
          type: [string, 'null']
          description: Cursor for the next page; null when no more pages.
        limit: { type: integer }
        count: { type: integer }
      required: [limit, count]
...
    Vertical:
      type: string
      description: Extensible enum; validate against the live set.
      enum: [angrosist, palletclearance, skalyou]
    Intent:
      type: string
      description: Extensible enum.
      enum: [buy, sell, market_entry]
    LeadStatus:
      type: string
      description: Extensible lead/offer state machine.
      enum:
        - new
        - qualifying
        - needs_human
        - qualified
        - offer_requested
        - offer_sent
        - negotiation
        - won
        - lost
    CompanyRole:
      type: string
      enum:
        - distributor
        - importer
        - wholesaler
        - retailer
        - horeca
        - processor
        - producer
        - buyer
        - seller
    LeadSummary:
      type: object
      description: List-view projection (mirrors domain.LeadSummary).
      properties:
        id: { type: string }
        status: { type: string }
        company_name: { type: string }
        cui: { type: string }
        product_name: { type: string }
        quantity:
          type: [number, 'null']
        unit: { type: string }
        delivery_location: { type: string }
        created_at: { type: string, format: date-time }
        vertical: { type: string }
        assigned_to: { type: [string, 'null'] }
        needs_human: { type: boolean }
        offer_value: { type: [number, 'null'] }
        offer_note: { type: string }
      required: [id, status, company_name, cui, product_name, quantity, unit, delivery_location, created_at, vertical, assigned_to, needs_human, offer_value]
    LeadListPage:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/LeadSummary' }
        page: { $ref: '#/components/schemas/PageInfo' }
      required: [data, page]
    HandoffItem:
      type: object
      description: One entry in the human-handoff queue (mirrors domain.HandoffItem).
      properties:
        id: { type: string }
        status: { type: string }
        vertical: { type: string }
        company_name: { type: string }
        product_name: { type: string }
        assigned_to: { type: [string, 'null'] }
        last_message: { type: string }
        created_at: { type: string, format: date-time }
      required: [id, status, vertical, company_name, product_name, assigned_to, last_message, created_at]
    TranscriptMessage:
      type: object
      properties:
        id: { type: string }
        conversation_id: { type: string }
        role: { type: string, examples: [user, assistant, tool] }
        content: { type: string }
        tool_calls:
          type: string
          description: Base64-encoded JSON array of tool-call parts.
        tool_call_id: { type: string }
        created_at: { type: string, format: date-time }
      required: [id, role, content, created_at]
    LeadDetail:
      allOf:
        - $ref: '#/components/schemas/LeadSummary'
        - type: object
          properties:
            address: { type: string }
            county: { type: string }
            phone: { type: string }
            email: { type: string }
            vertical: { $ref: '#/components/schemas/Vertical' }
            intent: { $ref: '#/components/schemas/Intent' }
            summary: { type: string }
            needs_human: { type: boolean }
            bot_active: { type: boolean }
            assigned_to: { type: [string, 'null'] }
            transcript:
              type: array
              items: { $ref: '#/components/schemas/TranscriptMessage' }
            sourcing_request:
              anyOf:
                - $ref: '#/components/schemas/SourcingRequest'
                - type: 'null'
            listing:
              anyOf:
                - $ref: '#/components/schemas/Listing'
                - type: 'null'
            market_entry_request:
              anyOf:
                - $ref: '#/components/schemas/MarketEntryRequest'
                - type: 'null'
          required: [address, county, phone, email, transcript]
    SourcingRequest:
      type: object
      description: Angrosist buyer demand (DATA_MODEL sourcing_requests).
      properties:
        lead_id: { type: string }
        category_id: { type: [string, 'null'] }
        product: { type: string }
        quantity: { type: [number, 'null'] }
        delivery_location: { type: string }
        recurring: { type: boolean }
        deadline: { type: [string, 'null'], format: date }
        budget: { type: [number, 'null'] }
      required: [lead_id, product]
    Listing:
      type: object
      description: PalletClearance supply (DATA_MODEL listings).
      properties:
        id: { type: string }
        lead_id: { type: string }
        company_id: { type: string }
        category_id: { type: [string, 'null'] }
        stock_type: { type: string }
        quantity: { type: [number, 'null'] }
        location: { type: string }
        country: { type: string }
        expiry: { type: [string, 'null'], format: date }
        target_price: { type: [number, 'null'] }
        confidential: { type: boolean }
        status: { type: string }
      required: [id, lead_id, status]
    MarketEntryRequest:
      type: object
      description: '[P2] SkalYou market-entry request (DATA_MODEL market_entry_requests).'
      properties:
        lead_id: { type: string }
        company_id: { type: string }
        category_id: { type: [string, 'null'] }
        partner_types:
          type: array
          items: { $ref: '#/components/schemas/CompanyRole' }
        has_permanent_stock: { type: boolean }
        active_markets:
          type: array
          items: { type: string, minLength: 2, maxLength: 2 }
        target_market: { type: string }
        website: { type: [string, 'null'], format: uri }
        objective: { type: string }
        specialist_type: { type: [string, 'null'] }
      required: [lead_id]
    Company:
      type: object
      properties:
        id: { type: string }
        country: { type: string, minLength: 2, maxLength: 2 }
        reg_no: { type: string, description: 'CUI for RO.' }
        name: { type: string }
        vat_status: { type: [string, 'null'] }
        caen: { type: [string, 'null'] }
        address: { type: [string, 'null'] }
        roles:
          type: array
          items: { $ref: '#/components/schemas/CompanyRole' }
        created_at: { type: string, format: date-time }
      required: [id, country, reg_no, name]
    CompanyDetail:
      allOf:
        - $ref: '#/components/schemas/Company'
        - type: object
          properties:
            verification:
              anyOf:
                - type: object
                  properties:
                    source: { type: string, examples: [demoanaf] }
                    vat_status: { type: [string, 'null'] }
                    administrators:
                      type: array
                      items: { type: string }
                    checked_at: { type: string, format: date-time }
                - type: 'null'
            financials:
              type: array
              items:
                type: object
                properties:
                  year: { type: integer }
                  turnover: { type: [number, 'null'] }
            contacts:
              type: array
              items: { $ref: '#/components/schemas/Contact' }
            leads:
              type: array
              items: { $ref: '#/components/schemas/LeadSummary' }
    Contact:
      type: object
      properties:
        id: { type: string }
        company_id: { type: string }
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        language: { type: string, enum: [ro, en] }
      required: [id]
...
    Kpis:
      type: object
      description: >
        Dashboard aggregates (mirrors domain.KPIs). offers_sent = leads in
        offer_sent|negotiation|won; won = leads in won; qualified = leads at
        qualification or beyond (the conversion denominator); conversion_rate =
        won/qualified (0 when qualified=0); pipeline_value = sum(offer_value) over
        open (non-terminal) statuses.
      properties:
        offers_sent: { type: integer }
        won: { type: integer }
        qualified: { type: integer }
        conversion_rate: { type: number }
        pipeline_value: { type: number }
      required: [offers_sent, won, qualified, conversion_rate, pipeline_value]
...